chore: Rename card to widget component
