+++ src/components/widget/Calendar/components/EventsSection.tsx
