+++ src/components/widget/LibraryCapacities/components/LibraryOccupancySection/components/Library.tsx
