+++ src/components/widget/LibraryCapacities/components/LibraryOccupancySection/components/ProgressCircle.tsx
