+++ src/components/widget/LibraryCapacities/components/LibraryOccupancySection/constants.ts
