+++ src/components/widget/LibraryCapacities/components/LibraryOccupancySection/index.tsx
