+++ src/components/widget/styled.tsx
